比赛大致信息,比赛详细信息，比赛是否被邀请判断更新以及传数组获取比赛题目

+++ src/main/java/com/example/oj/mapper/ContestUserMapper.java
@@ -9,5 +9,5 @@ import org.apache.ibatis.annotations.Select;
 public interface ContestUserMapper extends BaseMapper<ContestUser> {
 
     @Select("select from where contest = #{contestId} and user_id = #{userId}")
-    ContestUser getByContestIdAndUserId(Integer contestId, Long userId);
+    ContestUser getByContestIdAndUserId(Long contestId, Long userId);
 }
